Add the packaging and initial setup for the Python registries package.
`registries` is a Python package that provides classes for managing
the registration and lookup of classes, objects, or other state, backed
by unique lookup attributes. It originally comes from Djblets, but is
being split out and modernized.

This initial change sets up the packaging infrastructure, testing files,
and the initial package files. It does not include the registry
functionality, which will come in an upcoming change.

Testing Done:
Successfully built a package.

Ran `pytest` and `tox` with the in-progress implementation change.

Reviewed at https://reviews.reviewboard.org/r/14013/