Add a README for Registries.
This gives a description of Registries and the Registry Pattern,
plus installation instructions and example usage. We also have
information on how to contribute to Registries, and links to some other
projects we maintain.

Testing Done:
Rendered in a GitHub-compatible Markdown renderer and checked for
rendering issues.

Also checked for spelling issues and bad links.

Ran the example code to make sure it worked as expected.

Reviewed at https://reviews.reviewboard.org/r/14015/