Add documentation for registries.
This adds comprehensive documentation on how to use the registries
module. It includes some of the information from the README, along with
a code reference and guides on working with the different aspects of the
registries.

Testing Done:
Built the docs. Read through it, and checked links.

Reviewed at https://reviews.reviewboard.org/r/14016/